add redisson lock for follow

+++ src/main/java/com/fzq/xiaopotato/service/impl/UserfollowServiceImpl.java
@@ -2,12 +2,15 @@ package com.fzq.xiaopotato.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.esotericsoftware.minlog.Log;
 import com.fzq.xiaopotato.common.ErrorCode;
 import com.fzq.xiaopotato.common.NotificationType;
 import com.fzq.xiaopotato.common.utils.SocketIOUtils;
 import com.fzq.xiaopotato.exception.BusinessException;
 import com.fzq.xiaopotato.model.dto.common.IdDTO;
 import com.fzq.xiaopotato.model.entity.Likes;
+import com.fzq.xiaopotato.model.entity.Saves;
+import com.fzq.xiaopotato.model.entity.UserPost;
 import com.fzq.xiaopotato.model.entity.Userfollow;
 import com.fzq.xiaopotato.model.vo.NotificationVO;
 import com.fzq.xiaopotato.model.vo.UserVO;
@@ -15,12 +18,15 @@ import com.fzq.xiaopotato.service.UserService;
 import com.fzq.xiaopotato.service.UserfollowService;
 import com.fzq.xiaopotato.mapper.UserfollowMapper;
 import jakarta.servlet.http.HttpServletRequest;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.concurrent.TimeUnit;
 
 import static com.fzq.xiaopotato.common.NotificationType.FOLLOW;
 
@@ -44,6 +50,9 @@ public class UserfollowServiceImpl extends ServiceImpl<UserfollowMapper, Userfol
     @Autowired
     private SocketIOUtils socketIOUtils;
 
+    @Autowired
+    private RedissonClient redissonClient;
+
     @Override
     public boolean followByUserId(IdDTO idDTO, HttpServletRequest request) {
         UserVO user = userService.getCurrentUser(request);
@@ -55,20 +64,41 @@ public class UserfollowServiceImpl extends ServiceImpl<UserfollowMapper, Userfol
         if (followedId == followerId) {
             throw new BusinessException(ErrorCode.FORBIDDEN, "Cannot follow yourself.");
         }
-        boolean isFollowed =  isFollowed(followerId, followedId);
-        if (!isFollowed) {
-            Userfollow userfollow = new Userfollow();
-            userfollow.setFollowerId(followerId);
-            userfollow.setFollowedId(followedId);
-            userfollowMapper.insert(userfollow);
-            sendFollowNotification(user, followedId);
-            return true;
-        } else {
-            QueryWrapper<Userfollow> queryWrapper = new QueryWrapper<>();
-            queryWrapper.eq("follower_id", followerId).eq("followed_id", followedId);
-            userfollowMapper.delete(queryWrapper);
-            return false;
+        String lockKey = "user:follow:" + followerId + ":" + followedId;
+        RLock lock = redissonClient.getLock(lockKey);
+        try {
+            while (true) {
+                if (lock.tryLock(0, -1, TimeUnit.MILLISECONDS)) {
+                    Log.info("getLock: " + Thread.currentThread().getId());
+                    boolean isFollowed =  isFollowed(followerId, followedId);
+                    if (!isFollowed) {
+                        Userfollow userfollow = new Userfollow();
+                        userfollow.setFollowerId(followerId);
+                        userfollow.setFollowedId(followedId);
+                        userfollowMapper.insert(userfollow);
+                        sendFollowNotification(user, followedId);
+                        return true;
+                    } else {
+                        QueryWrapper<Userfollow> queryWrapper = new QueryWrapper<>();
+                        queryWrapper.eq("follower_id", followerId).eq("followed_id", followedId);
+                        userfollowMapper.delete(queryWrapper);
+                        return false;
+                    }
+                }
+            }
+        } catch (InterruptedException e) {
+            log.error(Thread.currentThread().getId() + " get lock failed");
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR, "get lock failed");
+        } finally {
+            // can only release itself lock
+            if (lock.isHeldByCurrentThread()) {
+                lock.unlock();
+                Log.info("release lock: {}", String.valueOf(Thread.currentThread().getId()));
+            }
+
         }
+
+
     }
 
     private boolean isFollowed(long followerId, long followedId) {
